rename directory under "src/sections/evidence" from "OpenPedCanSomaticMutations" to "OpenPedCanSomaticAlterations"
