+++ src/sections/evidence/OpenPedCanSomaticAlterations/SomaticMutationQuery.gql
